+++ MultiRobotSimulator.WPF/Events/CanvasRedrawEvent.cs
@@ -1,4 +1,12 @@
 ﻿namespace MultiRobotSimulator.WPF.Events
 {
-    public class CanvasRedrawEvent { }
+    public class CanvasRedrawEvent
+    {
+        private readonly static NLog.Logger _logger = NLog.LogManager.GetCurrentClassLogger();
+
+        public CanvasRedrawEvent()
+        {
+            _logger.Debug("Canvas redraw event called");
+        }
+    }
 }